Se aplicó correctamente el filtrado para peliculas

+++ CinematicMultiverse/src/main/java/com/example/CinematicMultiverse/resenhia/dto/CreateReseniaRequest.java
@@ -6,7 +6,7 @@ import java.util.UUID;
 public record CreateReseniaRequest(
         UUID usuarioId,
         UUID peliculaId,
-        int puntuacion,
+        double puntuacion,
         String comentario,
         LocalDate fechaPublicacion
 ) {
